requirements.txt
# Azure Integration
azure-identity==1.19.0
openai==1.58.1

# PDF and Document Processing
pdfplumber==0.11.4
python-docx==1.1.2

# Data Models and Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Data Processing and Storage
pandas==2.2.3
openpyxl==3.1.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Configuration
python-dotenv==1.0.1

# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
gunicorn==23.0.0
python-jose[cryptography]==3.3.0

# Utilities
rapidfuzz==3.10.1
tiktoken==0.8.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
